Fix ./build.sh --pack to work on Linux
There's no csi.exe in the .NET 5 (or .NET Core) SDK.  So move away from
a csx script to a full-blown tool written in C#.

For ease of review, I have kept the original tool code unmodified (not
even indentation has been fixed). To work around build warnings/errors,
disable the warnings.

Also fix the case of path to the dll in the local nuget cache. It
doesn't matter on Windows, but most file systems on Linux are
case-sensitive.